Initial Frontend Heirarchy
Usage:

Install node -> https://nodejs.org/en
Restart PC
Open Project (Ideally in IntelliJ which should come free with GitHub student developer pack)
Open Terminal (Alt + F12 OR bottom right of screen)
Write in terminal -> npm install (If this step fails, Node is not installed correctly)
After installation write in terminal -> lerna run dev --parallel
OR
Click on package.json in the root and click the green run button next to "both"

Open preffered browser and go to http://localhost:3000

Heirarchy specifications:

The root package.tsx in app is the landing page

The layout.tsx is how nextjs knows where the pages go. Later on we can do stuff like configure light/dark mode and make it so that unless the user is logged in, they wont be able to access other pages.

components/ - Root folder containing all reusable UI components for the application

landingComponents/ - Contains components specifically used for landing pages of the application

pages/ - Houses page-level components that represent complete views

services/ - Contains service-related files for API interactions and data fetching can also contain stuff like User Context

theme/ - Stores theming-related components specific to MUI

tmpComponents/ - Components of the tmp page

tmp/ - Temporary folder containing a page layout that will be deleted

Other Notes:

The tsconfig configurates typescript

The package.json stores information on how the code is ran  using scripts and what dependencies {packages) the specific folder needs.

Lerna is being used to manage the monorepo, when new dependencies (packages) are installed, there will be 3 places to put them:
-------
---  The BACKEND package.json:
This package.json stores ONLY the packages that the backend will use.

--- The FRONTEND package.json:
This  package.json stores ONLY the packages that the frontend will use.

--- The ROOT package..json:
This package.json stores the packages that the frontend AND backend will use. This allows for simpler package.json
-----

NOTE: the way to install new dependencies is to run this in the terminal:

npm install <your dependency>

BUT this will install the dependency in the ROOT package,json
So, all you have to do is change directory in the terminal to the frontend folder or the backend folder if the package doesnt need to be in the root directory:

cd packages/(frontend || backend)